packet: fix race condition in packet headers parsing
If one thread (in e.g. block A) calls is_tcp() on a packet, then
parse_iphdr() is called. If the IP header was not parsed yet the first
action in parse_iphdr() is to set the flag "m_iphdr_parsed =
true;". This is done before assigning any valid value to any of the IP
header fields. If in just that moment the same packet is accessed from
a second thread (in e.g. block B), then m_iphdr_parsed is already true
and the IP header fields are not touched even they are not set
correctly yet. In this case block B sees uninitialized values which
unpredictable consequences.

This patch adds proper synchronization to the header parsing code.

Signed-off-by: Yuri Volchkov <yuri.volchkov@neclab.eu>